Small patch to fix the "linux_redux" Chromium build.
Trying to DEPS roll the previous patch, which updates the
library to 1.0.1c, doesn't work for the following reason:

- third_party/libjingle is now referencing DTLSv1_client_method
  and DTLSv1_server_method, which were missing from openssl.gyp
  (this affects both Android and linux_redux builds).

- The piii/k8 configuration files didn't disable CAST5 and
  Whirlpool like the Android one, and the corresponding source
  files are missing from openssl.gyp.

- The piii config now enables inline assembly, but the
  source repository is missing the assembly source for
  the ia32 OPENSSL_setup_cpuid() function.

- The implementation source file for SHA-0 is not listed
  in openssl.gyp to match the Android build, which does
  explicitely disable this weak hash.

This patch is used to fix these issues in the most straightforward
way. It is possible to address the issues by adding missing source
files, but I'd rather do this in the upstream Android external/openssl
tree first, to minimize the differences between them.

BUG=137071
TEST=Passed net_unittests on both linux_redux and Android builds.

Review URL: https://chromiumcodereview.appspot.com/11467016

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/openssl@171794 4ff67af0-8c30-449e-8e8b-ad334ec8d88c