#including the correct header for ssize_t.
AEAD-related code could not be compiled on Mac and Windows
because the proper header file was not included and
SSIZE_MAX macro was missing.
This patch is a sort of workaround to make them compiled on Mac and Windows.

Non goal: care about all platforms that does not have ssize_t.
See the code review for details.

Review URL: https://codereview.chromium.org/98403012

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/openssl@240240 4ff67af0-8c30-449e-8e8b-ad334ec8d88c